4.21

+++ part4/controllers/blogs.js
@@ -33,8 +33,24 @@ blogRouter.post('/', async (request, response) => {
 });
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  // await Blog.findByIdAndDelete(request.params.id);
+  // response.status(204).end();
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+  const user = await User.findById(decodedToken.id);
+  const chosenBlog = await Blog.findById(request.params.id);
+
+  if (chosenBlog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } else {
+    response
+      .status(403)
+      .json({ error: 'Only user who created this blog can delete.' });
+  }
 });
 
 blogRouter.put('/:id', async (request, response) => {
